feat(app): Mise en place d'un fragment pour les produits de la page d'accueil
