--- app/app/src/main/res/layout/activity_main.xml.orig
@@ -1,173 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <!-- Header -->
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/top_bar_layout"
-        android:layout_width="411dp"
-        android:layout_height="69dp"
-        android:background="@color/cyan_background"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.0">
-
-        <SearchView
-            android:id="@+id/searchView"
-            android:layout_width="249dp"
-            android:layout_height="43dp"
-            android:background="@color/white"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.12"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <Button
-            android:id="@+id/main_search_button"
-            android:layout_width="118dp"
-            android:layout_height="50dp"
-            android:background="@drawable/button_background"
-            android:elevation="4dp"
-            android:text="@string/search_button"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.97"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <!-- Shop content -->
-    <TextView
-        android:id="@+id/main_shop_textview"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:text="@string/shop_main_textview_text"
-        android:textSize="24sp"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toTopOf="@+id/bottom_bar_layout"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/top_bar_layout"
-        app:layout_constraintVertical_bias="0.0" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/main_shop_recyclerview"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toTopOf="@+id/main_products_textview"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/main_shop_textview"
-        tools:listitem="@layout/shop_main_view" />
-
-    <!-- Product Content -->
-    <TextView
-        android:id="@+id/main_products_textview"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:text="@string/product_main_textview_text"
-        android:textSize="24sp"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/main_shop_textview"
-        app:layout_constraintVertical_bias="0.398" />
-
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/main_product_fragmentcontainerview"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toTopOf="@+id/bottom_bar_layout"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-<<<<<<< HEAD
-        app:layout_constraintTop_toBottomOf="@+id/main_products_textview"
-        app:layout_constraintVertical_bias="0.5"
-        tools:listitem="@layout/product_main_view" />
-=======
-        app:layout_constraintTop_toBottomOf="@+id/main_products_textview"/>
->>>>>>> 8267747 (refactor(app): Mise en place d'une nouvelle classe et XML pour utiliser un fragment sur la page d'accueil)
-
-    <!-- Footer -->
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/bottom_bar_layout"
-        android:layout_width="0dp"
-        android:layout_height="70dp"
-        android:background="@color/cyan_background"
-        android:orientation="horizontal"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/top_bar_layout"
-        app:layout_constraintVertical_bias="1.0">
-
-        <ImageButton
-            android:id="@+id/main_user_details_imagebutton"
-            android:layout_width="48dp"
-            android:layout_height="48dp"
-            android:background="@drawable/round_button"
-            android:clickable="true"
-            android:contentDescription="@string/profil_button_decription"
-            android:elevation="4dp"
-            android:scaleType="centerInside"
-            android:tint="@color/white"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.498"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:srcCompat="@drawable/person_fill0_24"
-            tools:ignore="UseAppTint" />
-
-        <ImageButton
-            android:id="@+id/main_home_imagebutton"
-            android:layout_width="48dp"
-            android:layout_height="48dp"
-            android:background="@drawable/round_button"
-            android:clickable="true"
-            android:contentDescription="@string/home_button_description"
-            android:elevation="4dp"
-            android:scaleType="centerInside"
-            android:src="@drawable/home_fill0_24"
-            android:tint="@color/white"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.231"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:ignore="UseAppTint" />
-
-        <ImageButton
-            android:id="@+id/main_shop_cart_imagebutton"
-            android:layout_width="48dp"
-            android:layout_height="48dp"
-            android:background="@drawable/round_button"
-            android:clickable="true"
-            android:contentDescription="@string/cart_button_description"
-            android:elevation="4dp"
-            android:scaleType="centerInside"
-            android:src="@drawable/shopping_fill0_z24"
-            android:tint="@color/white"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.765"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:ignore="UseAppTint" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
