--- app/app/src/main/res/values/strings.xml.orig
@@ -1,69 +1,0 @@
-<resources>
-    <string name="app_name">SpeedyCart</string>
-    <string name="email_textview_login_screen_text">E-mail</string>
-    <string name="password_textview_login_screen_text">Mot de passe</string>
-    <string name="forgetpasswordink_tewtview_login_screen_text">Mot de passe oublié?</string>
-    <string name="login_button_login_screen_text">Se connecter</string>
-    <string name="signup_textview_login_screen_text">Pas encore de compte ?</string>
-    <string name="signuplink_textview_login_screen_text">Inscrivez-vous</string>
-    <string name="email_textview_signup_screen_text">Email</string>
-    <string name="password_textview_signup_screen_text">Mot de passe</string>
-    <string name="passwordretype_textview_signup_screen_text">Entrez à nouveau le mot de passe</string>
-    <string name="usertype_textview_signup_screen_text">Vous êtes :</string>
-    <string name="title_textview_signup_screen_text">Inscription</string>
-    <string name="button_client_signup_screen_text">S\'enregistrer en tant que client</string>
-    <string name="title_textview_userinfo_screen_text">Mon compte client</string>
-    <string name="lastname_textview_userinfo_screen_text">Nom :</string>
-    <string name="firstname_textview_userinfo_screen_text">Prénom :</string>
-    <string name="mail_textview_userinfo_screen_text">E-mail :</string>
-    <string name="modify_button_userinfo_screen_text">Modifier des informations</string>
-    <string name="delete_button_userinfo_screen_text">Supprimer le compte</string>
-    <string name="loading">Loading...</string>
-    <string name="error">Error</string>
-    <string name="search_button">Recherche</string>
-    <string name="shop_main_textview_text">Magasins disponibles</string>
-    <string name="product_main_textview_text">Produits disponibles</string>
-    <string name="button_shop_signup_screen_text">S\'enregistrer en tant que vendeur/magasin</string>
-    <string name="button_delivery_person_signup_screen_text">S\'enregistrer en tant que livreur</string>
-    <string name="title_client_signup_screen_textview_text">Inscription</string>
-    <string name="lastname_textview_client_signup_screen_text">Nom de famille</string>
-    <string name="firstname_textview_client_signup_screen_text">Prénom</string>
-    <string name="dob_textview_client_signup_screen_text">Date de naissance</string>
-    <string name="adress_number_textview_client_signup_screen_text">Numéro</string>
-    <string name="adress_name_textview_client_signup_screen_text">Nom d\'adresse</string>
-    <string name="postal_code_textview_client_signup_screen_text">Code postale</string>
-    <string name="city_textview_client_signup_screen_text">Commune</string>
-    <string name="create_button_client_signup_screen_text">Créer le compte</string>
-    <string name="dob_button_client_signup_screen_text">Choisir une date</string>
-    <string name="title_shop_signup_screen_textview_text">Inscription</string>
-    <string name="name_textview_shop_signup_screen_text">Nom du vendeur/magasin</string>
-    <string name="description_textview_shop_signup_screen_text">Description</string>
-    <string name="siret_textview_shop_signup_screen_text">Numéro de SIRET</string>
-    <string name="address_number_textview_shop_signup_screen_text">Numéro</string>
-    <string name="address_name_textview_shop_signup_screen_text">Nom d\'adresse</string>
-    <string name="postal_code_textview_shop_signup_screen_text">Code postale</string>
-    <string name="city_textview_shop_signup_screen_text">Commune</string>
-    <string name="create_button_shop_signup_screen_text">Créer le compte</string>
-    <string name="title_textview_deliveryperson_signup_screen_text">Inscription</string>
-    <string name="lastname_textview_deliveryperson_signup_screen_text">Nom de famille</string>
-    <string name="firstname_textview_deliveryperson_signup_screen_text">Prénom</string>
-    <string name="dob_textview_deliveryperson_signup_screen_text">Date de naissance</string>
-    <string name="dob_button_deliveryperson_signup_screen_text">Choisir une date</string>
-    <string name="adress_number_textview_deliveryperson_signup_screen_text">Numéro</string>
-    <string name="adress_name_textview_deliveryperson_signup_screen_text">Nom d\'adresse</string>
-    <string name="postal_code_textview_deliveryperson_signup_screen_text">Code postale</string>
-    <string name="city_textview_deliveryperson_signup_screen_text">Commune</string>
-    <string name="create_button_deliveryperson_signup_screen_text">Créer le compte</string>
-    <string name="vehicle_textview_deliveryperson_signup_screen_text">Véhicule</string>
-<<<<<<< HEAD
-    <string name="home_button_description">home button</string>
-    <string name="cart_button_description">cart button</string>
-    <string name="profil_button_decription">profil button</string>
-    <string name="shop_display_description_text">magasin de vente de proximité</string>
-    <string name="shop_display_address_text">34 rue de feuille</string>
-    <string name="shop_display_title_text">Intermarché</string>
-=======
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
->>>>>>> 8267747 (refactor(app): Mise en place d'une nouvelle classe et XML pour utiliser un fragment sur la page d'accueil)
-</resources>
